Split WindowsSDK class into a separate module and log MSVCSetupAPI errors.
