+++ lib/Utils/CMakeLists.txt
@@ -43,5 +43,5 @@ if ( MSVC )
   set_target_properties(clingUtils PROPERTIES
     COMPILE_FLAGS "/D_HAS_EXCEPTIONS=0")
   target_sources(clingUtils PRIVATE
-    "${CMAKE_CURRENT_SOURCE_DIR}/Windows/MSVCSetupApi.cpp")
+    "${CMAKE_CURRENT_SOURCE_DIR}/Windows/WindowsSDK.cpp")
 endif ()
